fix bug with absolute moves
